Refactor move to send back a player response
