+++ src/engine/player_response.rs
@@ -1,0 +1,7 @@
+use crate::models::{Piece, Player};
+
+pub struct PlayerResponse {
+	pub player: Player,
+	pub piece: Piece,
+	pub raw_response: String,
+}
