+++ src/models/mod.rs
@@ -1,10 +1,14 @@
 pub mod piece;
-pub mod plateau;
-pub mod player;
-pub mod point;
-
 pub use piece::Piece;
 pub use piece::PieceBag;
+
+pub mod plateau;
 pub use plateau::Plateau;
+
+pub mod player;
 pub use player::Player;
+
+pub mod point;
 pub use point::Point;
+
+mod constants;
